AlertHandler/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-common
azure-identity
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-network
requests